Nav-helper should be used in modal and should close modal on link click
