+++ src/modules/modal/routes.js
@@ -3,6 +3,7 @@ import { connect } from 'react-redux';
 
 import Modal from 'react-modal'
 import SettingsContainer from 'gComponents/users/settings/container.js'
+import NavHelper from 'gComponents/utility/NavHelper/index.js';
 import * as modalActions from 'gModules/modal/actions.js'
 
 const routes = [
@@ -44,13 +45,15 @@ export default class ModalRouter extends Component{
 
       const isOpen = true
       return(
-        <Modal
-          isOpen={isOpen}
-          onRequestClose={this.props.onClose}
-          style={customStyles}
-        >
-          {<item.component {...props} onClose={this.props.onClose}/>}
-        </Modal>
+          <Modal
+            isOpen={isOpen}
+            onRequestClose={this.props.onClose}
+            style={customStyles}
+          >
+            <NavHelper>
+                {<item.component {...props} onClose={this.props.onClose}/>}
+            </NavHelper>
+          </Modal>
       )
     }
   }
