--- src/routes/layouts/application/nav-helper.js
@@ -1,24 +1,0 @@
-import app from 'ampersand-app'
-import React from 'react'
-import localLinks from 'local-links'
-
-export default React.createClass({
-  displayName: 'NavHelper',
-
-  onClick (event) {
-    const pathname = localLinks.getLocalPathname(event)
-
-    if (pathname) {
-      event.preventDefault()
-      app.router.history.navigate(pathname)
-    }
-  },
-
-  render () {
-    return (
-      <div className='navHelper' {...this.props} onClick={this.onClick}>
-        {this.props.children}
-      </div>
-    )
-  }
-})
